Moved initMiddleware to utils and added line ending
