+++ music-journal/utils/initMiddleware.js
